+++ app/vmalert-tool/unittest/unittest_test.go
@@ -8,7 +8,7 @@ func TestUnitTest_Failure(t *testing.T) {
 	f := func(files []string) {
 		t.Helper()
 
-		failed := UnitTest(files, false, nil, "", "", "")
+		failed := UnitTest(files, false, nil, "", "", "", false)
 		if !failed {
 			t.Fatalf("expecting failed test")
 		}
@@ -23,7 +23,7 @@ func TestUnitTest_Success(t *testing.T) {
 	f := func(disableGroupLabel bool, files []string, externalLabels []string, externalURL, httpPort string) {
 		t.Helper()
 
-		failed := UnitTest(files, disableGroupLabel, externalLabels, externalURL, httpPort, "")
+		failed := UnitTest(files, disableGroupLabel, externalLabels, externalURL, httpPort, "", false)
 		if failed {
 			t.Fatalf("unexpected failed test")
 		}
@@ -31,6 +31,9 @@ func TestUnitTest_Success(t *testing.T) {
 
 	// run multi files with random http port
 	f(false, []string{"./testdata/test1.yaml", "./testdata/test2.yaml"}, []string{"cluster=prod"}, "http://grafana:3000", "")
+
+	// rule groups depending on each other across groups
+	f(false, []string{"./testdata/rule-deps-test.yaml"}, nil, "", "")
 
 	// disable group label
 	// template with null external values